Update Display.java
The option to mark nodes has been added. But there is no option to unmark the node in O.Zimmermann's code (if you see this, perhaps you could change it in your code).
Also, the color of the marker in dark mode has been changed from red to green.
I changed the settings window to a menu window because I want to do more with the current window.